requirements.txt
ollama
pyyaml
yfinance